separate SwiftUI widgets demos
