--- DemoDirectory/DemoDirectory/Demos/FilterList/FilterListSwiftUIDemoViewController.swift
@@ -1,15 +1,0 @@
-//
-//  FilterListSwiftUIDemoViewController.swift
-//  DemoDirectory
-//
-//  Created by Vladislav Fitc on 30/06/2021.
-//  Copyright © 2021 Algolia. All rights reserved.
-//
-
-import Foundation
-
-class FilterListSwiftUIDemoController {
-  
-  
-  
-}
