+++ DemoDirectory/DemoDirectory/Demos/Stats/StatsDemoViewController.swift
@@ -95,22 +95,3 @@ private extension StatsDemoViewController {
   
 }
 
-import SwiftUI
-
-struct StatsDemoSwiftUI: PreviewProvider {
-    
-  struct ContentView: View {
-    
-    @ObservedObject var statsController: StatsTextObservableController
-    
-    var body: some View {
-      Text(statsController.stats)
-    }
-    
-  }
-  
-  static var previews: some View {
-    ContentView(statsController: .init(stats: "Hello World!"))
-  }
-  
-}
